Reworked Client ID handling/management
* RiakConnectionUsageManager no longer auto generates the client ID
  per call. It now takes the client ID as part of its interface and
  sets it on the connection prior to the call being made.
* RiakClient now contains the client ID and that is passed along with
  each call to the UseConnection() function so that it is set on the
  connection before the call is made.
* RiakClient will automatically create a client ID based on the MAC
  address of the first active NIC found on the machine. This is
  intended to avoid the need for Rich Client applications to worry
  about the Client ID at all. In the web world, the clients will
  need to be smarter in the way this is used given that the RiakClient
  is disposed in each connection. RiakClients could potentially be
  stored in the user's session so that the client lives alongside
  the user and avoids the need to worry, but this might not be the
  option that devs want to go with. They can set their own ID when
  the RiakClient instance is created. It should in some way be tied
  to the one user. The goal of this is to reduce the guaranteed bloat
  of vclocks.

More work will be done on this as having to set the client ID prior
to each call isn't exactly the most optimal, yet it seems to be the
only option when it comes to dealing with pooling of connections.

I'll be emailing the Riak mailing list soon to talk about the options
because, after talking to Russell, I agree with his suggested approach
of having the ClientID passed in with each of the calls as an extra
parameter on the PB objects. This avoids the need to set the clientID
on each of the connections when they're acquired. 4 bytes of overhead
per call is negligable and way easier to manage.