Fix #21 -- added the GetServerInfo() API call
Returns a strongly typed object that contains the name of the node
that the call was executed on and the version of Riak that the node
is running.

This assumes that all of the nodes in the cluster are running the same
version of Riak. That is, if this isn't the case, then the user will
only get the details of the one node in the cluster they just happened
to execute the call against.

If this level of functionality doesn't suffice (which I would think it
does in most cases) then we can perhaps look to extend it down the track
with the ability to call against every single node in the cluster and
get a list of entries back rather than a single entry (overkill imho).


Former-commit-id: f28f28203e837edd2780222a196020be58d6c381