Moved KeyFilters to spearate Models.MapReduce namespace. Added RiakFluentKeyFilter class. Created RiakMapReduceTests to localize the unit tests for the fluent key filters and future MR tests. Added methods to RiakMapReduceQuery to use the RiakFluentKeyFilter class. Updated KeyFilter internal data structures to use read only Tuples. Updated RiakInput classes with a substantial re-write to fix issues with serializing the objects to JSON.
Created RiakMapReduceTests to localize MR specific tests.

Updated tests and objects to support fluent key filters

Added additional tests for FluentKeyFilter
Modified FluentKeyFilter to use a single IRiakKeyFilterToken instead of List<IRiakKeyFilterToken>
Updated RiakBucketInput to fix issue with converting Inputs property to JSON.
Updated Compile to include property for inputs.

Changed test case to use constructor config section.
Switched FluentKeyFilter back to using a List of KeyFilters - after checking the KeyFilter docs I realize that this is possible using a combination of Predicates and Key Filters
Modified KeyFilter WriteArgumentsAsArray to remove wrapping nested args in arrays. This is already being done by some other part of the serialization code so... yeah, don't need it.

Updated KeyFilters and Inputs to use Tuples and other assorted fancy data structures.
Made a variety of API changes to remove a considerable amount of suck and replace it with an equal amount of win.
Updated unit tests to use the aforementioned changes to the API.

Modified RiakInput subclass ToJsonString methods and RiakMapReduceQuery.Compile

Moved responsibility for figuring out how to turn a RiakPhaseInput subclass into JSON out to the subclass implementations. Which is as it should be, really.

Cleaned up KeyFilter code to use readonly tuples and Tuple.Create syntax for easier code readability.
Made minor changes to And, Between, Not, and Or to allow for Left and Right values (or just Arg for Not) to contain a List<IRiakKeyFilter> giving us truly bitchin' KeyFilter functionality.
Commented out some garbage unit tests before removing them completely in the final round of code pushes.